Refactor E2E error handling tests
- Extract verifyErrorDisplay and verifyNoError to shared test helpers
- Move connection error tests to separate spec file
- Rename error-handling.spec.ts to user-rejection-errors.spec.ts for clarity
- Simplify connection error tests to focus on testable scenarios
