+++ apps/e2e/src/user-rejection-errors.spec.ts
@@ -1,28 +1,12 @@
 import { testWithSynpress } from '@synthetixio/synpress';
 import { MetaMask, metaMaskFixtures } from '@synthetixio/synpress/playwright';
-import { Page } from '@playwright/test';
 import basicSetup from '../test/wallet-setup/basic.setup';
+import { verifyErrorDisplay, verifyNoError } from './utils/test-helpers';
 
 const test = testWithSynpress(metaMaskFixtures(basicSetup));
 const { expect } = test;
 
-test.describe('Error Handling', () => {
-  // Helper function to verify error display
-  const verifyErrorDisplay = async (
-    page: Page,
-    expectedError: string
-  ): Promise<void> => {
-    const errorElement = page.locator('[data-testid="error-message"]');
-    await expect(errorElement).toBeVisible({ timeout: 5000 });
-    await expect(errorElement).toHaveText(expectedError);
-  };
-
-  // Helper function to verify no error is displayed
-  const verifyNoError = async (page: Page): Promise<void> => {
-    const errorElement = page.locator('[data-testid="error-message"]');
-    await expect(errorElement).not.toBeVisible();
-  };
-
+test.describe('User Rejection Errors', () => {
   test.describe('Transaction Rejection Errors', () => {
     test('should show error when user rejects swap transaction', async ({
       context,
@@ -376,281 +360,6 @@ test.describe('Error Handling', () => {
 
       // Verify error is cleared
       await verifyNoError(page);
-    });
-  });
-
-  test.describe('Connection Errors', () => {
-    test('should show error when wallet connection is lost during transaction', async ({
-      context,
-      page,
-      metamaskPage,
-      extensionId,
-    }) => {
-      const metamask = new MetaMask(
-        context,
-        metamaskPage,
-        'Tester@1234',
-        extensionId
-      );
-
-      // Setup: Connect wallet first
-      await page.goto('/');
-      await page
-        .locator('button')
-        .filter({ hasText: /connect/i })
-        .first()
-        .click();
-      await metamask.connectToDapp();
-
-      // Wait for connection
-      await expect(
-        page.getByRole('heading', { name: 'Simple AMM DApp' })
-      ).toBeVisible();
-
-      // Start a swap
-      const swapSection = page
-        .getByRole('heading', { name: 'Swap' })
-        .locator('../..');
-      const swapInput = swapSection.getByPlaceholder('SIMP → ETH');
-      await swapInput.fill('1');
-
-      const swapButton = page.getByRole('button', {
-        name: 'Swap SIMP for ETH',
-      });
-      await swapButton.click();
-
-      // Wait for the transaction prompt to appear and then simulate connection loss
-      await page.waitForTimeout(2000);
-
-      // Simulate network/connection error by closing MetaMask pages
-      const metamaskPages = context
-        .pages()
-        .filter(
-          (p) =>
-            p.url().includes('chrome-extension') &&
-            p.url().includes('notification')
-        );
-      for (const metamaskPage of metamaskPages) {
-        await metamaskPage.close();
-      }
-
-      // Wait and verify error is shown (should be a connection/network error)
-      await page.waitForTimeout(5000);
-
-      // The error should be some form of connection failure
-      const errorElement = page.locator('[data-testid="error-message"]');
-      await expect(errorElement).toBeVisible({ timeout: 10000 });
-
-      // Verify it contains an error message (exact message may vary)
-      const errorText = await errorElement.textContent();
-      expect(errorText).toContain('failed:');
-    });
-
-    test('should handle network disconnection gracefully', async ({
-      context,
-      page,
-      metamaskPage,
-      extensionId,
-    }) => {
-      const metamask = new MetaMask(
-        context,
-        metamaskPage,
-        'Tester@1234',
-        extensionId
-      );
-
-      // Setup: Connect wallet first
-      await page.goto('/');
-      await page
-        .locator('button')
-        .filter({ hasText: /connect/i })
-        .first()
-        .click();
-      await metamask.connectToDapp();
-
-      // Wait for connection
-      await expect(
-        page.getByRole('heading', { name: 'Simple AMM DApp' })
-      ).toBeVisible();
-
-      // Simulate going offline by intercepting network requests
-      await context.route('http://localhost:8545', (route) => {
-        route.abort('internetDisconnected');
-      });
-
-      // Try to perform a swap which should fail due to network
-      const swapSection = page
-        .getByRole('heading', { name: 'Swap' })
-        .locator('../..');
-      const swapInput = swapSection.getByPlaceholder('SIMP → ETH');
-      await swapInput.fill('1');
-
-      const swapButton = page.getByRole('button', {
-        name: 'Swap SIMP for ETH',
-      });
-      await swapButton.click();
-
-      // Should show network/connection error
-      const errorElement = page.locator('[data-testid="error-message"]');
-      await expect(errorElement).toBeVisible({ timeout: 15000 });
-
-      const errorText = await errorElement.textContent();
-      expect(errorText).toContain('Swap failed:');
-      // Network errors often contain terms like "fetch", "network", or "connection"
-      expect(errorText?.toLowerCase()).toMatch(
-        /fetch|network|connection|request/
-      );
-    });
-
-    test('should handle RPC endpoint failure', async ({
-      context,
-      page,
-      metamaskPage,
-      extensionId,
-    }) => {
-      const metamask = new MetaMask(
-        context,
-        metamaskPage,
-        'Tester@1234',
-        extensionId
-      );
-
-      // Setup: Connect wallet first
-      await page.goto('/');
-      await page
-        .locator('button')
-        .filter({ hasText: /connect/i })
-        .first()
-        .click();
-      await metamask.connectToDapp();
-
-      // Wait for connection
-      await expect(
-        page.getByRole('heading', { name: 'Simple AMM DApp' })
-      ).toBeVisible();
-
-      // Simulate RPC errors by returning error responses
-      await context.route('http://localhost:8545', (route) => {
-        route.fulfill({
-          status: 500,
-          contentType: 'application/json',
-          body: JSON.stringify({
-            jsonrpc: '2.0',
-            id: 1,
-            error: {
-              code: -32603,
-              message: 'Internal server error',
-            },
-          }),
-        });
-      });
-
-      // Try to perform a swap which should fail due to RPC error
-      const swapSection = page
-        .getByRole('heading', { name: 'Swap' })
-        .locator('../..');
-      const swapInput = swapSection.getByPlaceholder('SIMP → ETH');
-      await swapInput.fill('1');
-
-      const swapButton = page.getByRole('button', {
-        name: 'Swap SIMP for ETH',
-      });
-      await swapButton.click();
-
-      // Should show RPC error
-      const errorElement = page.locator('[data-testid="error-message"]');
-      await expect(errorElement).toBeVisible({ timeout: 15000 });
-
-      const errorText = await errorElement.textContent();
-      expect(errorText).toContain('Swap failed:');
-    });
-
-    test('should handle wallet disconnection during transaction', async ({
-      context,
-      page,
-      metamaskPage,
-      extensionId,
-    }) => {
-      const metamask = new MetaMask(
-        context,
-        metamaskPage,
-        'Tester@1234',
-        extensionId
-      );
-
-      // Setup: Connect wallet first
-      await page.goto('/');
-      await page
-        .locator('button')
-        .filter({ hasText: /connect/i })
-        .first()
-        .click();
-      await metamask.connectToDapp();
-
-      // Wait for connection
-      await expect(
-        page.getByRole('heading', { name: 'Simple AMM DApp' })
-      ).toBeVisible();
-
-      // Start a transaction
-      const swapSection = page
-        .getByRole('heading', { name: 'Swap' })
-        .locator('../..');
-      const swapInput = swapSection.getByPlaceholder('SIMP → ETH');
-      await swapInput.fill('1');
-
-      const swapButton = page.getByRole('button', {
-        name: 'Swap SIMP for ETH',
-      });
-      await swapButton.click();
-
-      // Wait for transaction prompt
-      await page.waitForTimeout(3000);
-
-      // Simulate wallet disconnection by navigating to MetaMask and disconnecting
-      await metamaskPage.bringToFront();
-
-      // Try to disconnect by going to connected sites if available
-      try {
-        // This might vary based on MetaMask version, so we'll be flexible
-        await metamaskPage.click('[data-testid="account-menu-icon"]', {
-          timeout: 5000,
-        });
-        await page.waitForTimeout(1000);
-
-        // Look for connected sites or similar option
-        const connectedSitesSelector =
-          'text=/Connected sites|Sites|Connections/i';
-        if ((await metamaskPage.locator(connectedSitesSelector).count()) > 0) {
-          await metamaskPage.click(connectedSitesSelector);
-          await page.waitForTimeout(1000);
-
-          // Try to disconnect
-          const disconnectSelector = 'text=/Disconnect|Remove/i';
-          if ((await metamaskPage.locator(disconnectSelector).count()) > 0) {
-            await metamaskPage.click(disconnectSelector);
-          }
-        }
-      } catch {
-        // If the above doesn't work, just close the notification window
-        const notificationPages = context
-          .pages()
-          .filter((p) => p.url().includes('notification.html'));
-        for (const notPage of notificationPages) {
-          await notPage.close();
-        }
-      }
-
-      // Go back to main page and verify error handling
-      await page.bringToFront();
-      await page.waitForTimeout(3000);
-
-      // Should show some form of wallet connection error
-      const errorElement = page.locator('[data-testid="error-message"]');
-      await expect(errorElement).toBeVisible({ timeout: 15000 });
-
-      const errorText = await errorElement.textContent();
-      expect(errorText).toContain('failed:');
     });
   });
 });
